requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
requests==2.32.3
apscheduler==3.10.4
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
